# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个基于Go和Gin的博客系统API文档",
        "title": "Bmtdblog API",
        "termsOfService": "http://localhost:8090",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8090",
            "email": "3138910969@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/",
    "paths": {
...
        "/admin/page/create": {
            "post": {
                "description": "创建新的博客页面",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "页面管理"
                ],
                "summary": "创建新页面",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页面标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页面内容",
                        "name": "body",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "是否发布(on表示发布)",
                        "name": "isPublished",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建失败的页面",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "重定向到页面管理页面",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/page/{id}/delete": {
            "post": {
                "description": "删除指定ID的页面",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "页面管理"
                ],
                "summary": "删除页面",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页面ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"succeed\":false,\"message\":\"错误信息\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/page/{id}/edit": {
            "get": {
                "description": "管理后台编辑页面的表单页面",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "页面管理"
                ],
                "summary": "编辑页面表单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页面ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "编辑页面表单",
                        "schema": {
                            "type": "html"
                        }
                    },
                    "404": {
                        "description": "页面不存在",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "更新指定ID的页面内容",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "页面管理"
                ],
                "summary": "更新页面",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页面ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页面标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页面内容",
                        "name": "body",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "是否发布(on表示发布)",
                        "name": "isPublished",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "301": {
                        "description": "更新成功，重定向到页面列表",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/admin/post/create": {
            "post": {
                "description": "创建新的博客文章并添加标签",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "内容管理"
                ],
                "summary": "创建新文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文章内容",
                        "name": "body",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文章标签ID，逗号分隔",
                        "name": "tags",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "是否发布(on表示发布)",
                        "name": "isPublished",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建失败的页面",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "重定向到文章管理页面",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/admin/post/{id}/delete": {
            "post": {
                "description": "删除指定ID的文章及其相关标签",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"succeed\":false,\"message\":\"错误信息\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/admin/post/{id}/publish": {
            "post": {
                "description": "切换指定ID文章的发布状态（发布/下架）",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "切换文章发布状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"succeed\":false,\"message\":\"错误信息\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/page/{id}": {
            "get": {
                "description": "根据ID获取静态页面的详细内容",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "内容浏览"
                ],
                "summary": "页面详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页面ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "页面详情页面",
                        "schema": {
                            "type": "html"
                        }
                    },
                    "404": {
                        "description": "页面不存在或未发布",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/visitor/comment/{id}/delete": {
            "post": {
                "description": "删除自己发布的评论",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论管理"
                ],
                "summary": "删除评论",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "评论ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"succeed\":false,\"message\":\"错误信息\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}